requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cors==1.0.1
cryptography==36.0.1
dnspython==2.1.0
filelock==3.4.0
Flask==2.0.2
Flask-Cors==3.0.10
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
PyJWT==2.3.0
pymongo==4.0.1
PySocks==1.7.1
python-dotenv==0.19.2
requests==2.26.0
requests-file==1.5.1
six==1.16.0
tldextract==3.1.2
urllib3==1.26.7
Werkzeug==2.0.2
zope.event==4.5.0
zope.interface==5.4.0
